# src/docs/swagger.yaml
definitions:
...
  dto.CreatePropertyCategoryRequest:
    properties:
      icon:
        maxLength: 1000
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
  dto.CreatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      unit:
        maxLength: 15
        type: string
    required:
    - categoryId
    - name
    type: object
...
  dto.UpdatePropertyCategoryRequest:
    properties:
      icon:
        type: string
      name:
        type: string
    type: object
  dto.UpdatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        type: string
      unit:
        maxLength: 15
        type: string
    type: object
  filter.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
  filter.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/filter.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/filter.Sort'
        type: array
    type: object
  filter.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/validation.ValidationError'
        type: array
    type: object
  helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/properties/:
    post:
      consumes:
      - application/json
      description: Create a Property
      parameters:
      - description: Create a Property
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePropertyRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a Property
      tags:
      - Properties
  /v1/properties/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a Property
      tags:
      - Properties
    get:
      consumes:
      - application/json
      description: Get a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Property
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Update a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Property
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePropertyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a Property
      tags:
      - Properties
  /v1/properties/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Properties
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/filter.PaginationInputWithFilter'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Properties
      tags:
      - Properties
  /v1/property-categories/:
    post:
      consumes:
      - application/json
      description: Create a PropertyCategory
      parameters:
      - description: Create a PropertyCategory
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePropertyCategoryRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a PropertyCategory
      tags:
      - PropertyCategories
  /v1/property-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a PropertyCategory
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a PropertyCategory
      tags:
      - PropertyCategories
    get:
      consumes:
      - application/json
      description: Get a PropertyCategory
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a PropertyCategory
      tags:
      - PropertyCategories
    put:
      consumes:
      - application/json
      description: Update a PropertyCategory
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a PropertyCategory
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePropertyCategoryRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a PropertyCategory
      tags:
      - PropertyCategories
  /v1/property-categories/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get PropertyCategories
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/filter.PaginationInputWithFilter'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get PropertyCategories
      tags:
      - PropertyCategories
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"